split email mailbox processor into two seperate services
